# C# - Optimization and clean up