--- a/csharp/MonkeyLang.PerformanceTests/MonkeyLang.PerformanceTests.csproj
+++ b/csharp/MonkeyLang.PerformanceTests/MonkeyLang.PerformanceTests.csproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net7.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.13.5" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MonkeyLang\MonkeyLang.csproj" />
+  </ItemGroup>
+
+</Project>