--- a/csharp/MonkeyLang.PerformanceTests/Program.cs
+++ b/csharp/MonkeyLang.PerformanceTests/Program.cs
@@ -1,0 +1,4 @@
+﻿using BenchmarkDotNet.Running;
+using MonkeyLang.PerformanceTests;
+
+BenchmarkRunner.Run<LexerTests>();