--- a/csharp/MonkeyLang.Tests/MonkeyLang.Tests.csproj
+++ b/csharp/MonkeyLang.Tests/MonkeyLang.Tests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
         <TargetFramework>net7.0</TargetFramework>
@@ -18,7 +18,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\lexer\lexer.csproj" />
+      <ProjectReference Include="..\MonkeyLang\MonkeyLang.csproj" />
     </ItemGroup>
 
 </Project>