--- a/csharp/MonkeyLang.sln
+++ b/csharp/MonkeyLang.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MonkeyLang", "MonkeyLang\Mo
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MonkeyLang.Tests", "MonkeyLang.Tests\MonkeyLang.Tests.csproj", "{35C8E3B4-7A74-4B2B-8923-BEB6128DE98F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MonkeyLang.PerformanceTests", "MonkeyLang.PerformanceTests\MonkeyLang.PerformanceTests.csproj", "{A4DFA55A-B9DB-4A40-8075-ECFB664E069A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{35C8E3B4-7A74-4B2B-8923-BEB6128DE98F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{35C8E3B4-7A74-4B2B-8923-BEB6128DE98F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{35C8E3B4-7A74-4B2B-8923-BEB6128DE98F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A4DFA55A-B9DB-4A40-8075-ECFB664E069A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A4DFA55A-B9DB-4A40-8075-ECFB664E069A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A4DFA55A-B9DB-4A40-8075-ECFB664E069A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A4DFA55A-B9DB-4A40-8075-ECFB664E069A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE