--- a/csharp/MonkeyLang/Dockerfile
+++ b/csharp/MonkeyLang/Dockerfile
@@ -2,9 +2,9 @@
 WORKDIR /src
 COPY . .
 RUN dotnet restore
-RUN dotnet publish "lexer.csproj" -c Release -o /app/publish
+RUN dotnet publish "MonkeyLang.csproj" -c Release -o /app/publish
 
 FROM mcr.microsoft.com/dotnet/runtime:7.0 AS final
 WORKDIR /app
 COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "lexer.dll"]
+ENTRYPOINT ["dotnet", "MonkeyLang.dll"]