--- a/csharp/MonkeyLang/Lexer.cs
+++ b/csharp/MonkeyLang/Lexer.cs
@@ -1,4 +1,4 @@
-﻿namespace monkey;
+﻿namespace MonkeyLang;
 
 public struct Lexer
 {