--- a/csharp/MonkeyLang/Program.cs
+++ b/csharp/MonkeyLang/Program.cs
@@ -1,4 +1,4 @@
-﻿using monkey;
+﻿using MonkeyLang;
 
 var user = Environment.UserName;
 
@@ -8,18 +8,18 @@ Console.WriteLine("Feel free to type in commands");
 const string PROMPT = ">> ";
 
 while (true)
-{ 
+{
     Console.Write(PROMPT);
     var line = Console.ReadLine();
 
-    if(line is null)
+    if (line is null)
     {
         break;
     }
 
     var lexer = new Lexer(line);
 
-    foreach(var token in lexer.ParseTokens())
+    foreach (var token in lexer.ParseTokens())
     {
         Console.WriteLine(token);
     }