--- a/csharp/MonkeyLang/Token.cs
+++ b/csharp/MonkeyLang/Token.cs
@@ -1,4 +1,4 @@
-﻿namespace monkey;
+﻿namespace MonkeyLang;
 
 public enum Token
 {