--- a/csharp/lexer.tests/LexerTest.cs
+++ b/csharp/lexer.tests/LexerTest.cs
@@ -1,3 +1,5 @@
+using monkey;
+
 namespace lexer.tests;
 
 public class Tests