This pull request is built upon the foundation set by @redducks100 PR #79. I focused on improving the efficiency of the `Lexer` and improved readability of the code base. I've added an performance test project. The results of the performance tests can be found below:

``` ini

BenchmarkDotNet=v0.13.5, OS=Windows 10 (10.0.19045.2965/22H2/2022Update)
Intel Core i7-7700 CPU 3.60GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET SDK=7.0.400-preview.23225.8
  [Host]     : .NET 7.0.5 (7.0.523.17405), X64 RyuJIT AVX2
  DefaultJob : .NET 7.0.5 (7.0.523.17405), X64 RyuJIT AVX2


```

**redducks100** implementation:
|        Method |      Mean |     Error |    StdDev |       Min |       Max |   Gen0 | Allocated |
|-------------- |----------:|----------:|----------:|----------:|----------:|-------:|----------:|
| SmallCodeTest |  2.157 μs | 0.0227 μs | 0.0212 μs |  2.123 μs |  2.181 μs | 0.1450 |     616 B |
| LargeCodeTest | 55.294 μs | 0.5759 μs | 0.5387 μs | 54.426 μs | 56.098 μs | 3.6621 |   15400 B |

My implementation: *(44.15% faster)*
|        Method |      Mean |     Error |    StdDev |       Min |       Max |   Gen0 | Allocated |
|-------------- |----------:|----------:|----------:|----------:|----------:|-------:|----------:|
| SmallCodeTest |  1.250 μs | 0.0042 μs | 0.0040 μs |  1.243 μs |  1.256 μs | 0.1583 |     664 B |
| LargeCodeTest | 30.897 μs | 0.3221 μs | 0.3013 μs | 30.314 μs | 31.177 μs | 3.9673 |   16600 B |

The small code test includes all token types within 16 lines of code, the large test is the same code repeated 25 times resulting in a total of 448 lines of code.

I tried using a `ReadOnlyMemory<char>` or `Range` instead of `string` for the `Literal` property in `Token`. Of course it resulted in no allocations, but the performance had to suffer instead *(2x slower)*, i'm not sure why this happened, I guess strings are just too optimized.