@vhladko 2023-06-04
Hello, thanks for your effort, the only one problem is that we already have csharp and csharp2 folders, and @redducks100 still need to create a new one as a separate folder, you could wait for him or maybe work together and merge your updated version as csharp 3, also don’t forget to add you as a Codeowner :)