@redducks100 2023-06-04
We have decided to push another pr similar to mine @Abooow (check csharp2 implementation), if you want you could add another folder (csharp3) but from what I can see the implementation is really similar to what we already have.