@redducks100 2023-06-04
> @redducks100 thanks for clarification, so can we then also close your PR ? #79

@leo-costa wanted to benchmark my PR and asked not to close it yet, we could move it to draft tho