@Abooow 2023-06-04
This PR was intended to be what the `csharp2` folder currently is, i guess i was too slow. But it's no problem since they are quite similar anyway.
I don't think we need a `csharp3` folder, having one implementation focused on performance and another on language features should be sufficient. I'm willing to help and contribute in any of the existing implementations. :)