@leo-costa 2023-06-04
Here are some benchmarks that I promised @Abooow @redducks100 @TomasDurica ! Have good weekend 
<img width="993" alt="image" src="https://github.com/ThePrimeagen/ts-rust-zig-deez/assets/32940986/7871e159-abe8-4932-a3ef-bd54278c0739">